32bit compat

+++ pandas/tests/arrays/masked_shared.py
@@ -148,7 +148,7 @@ class NumericOps:
     def test_mul_td64_array(self, dtype):
         # GH#45622
         arr = pd.array([1, 2, pd.NA], dtype=dtype)
-        other = np.arange(3).view("m8[ns]")
+        other = np.arange(3, dtype=np.int64).view("m8[ns]")
 
         result = arr * other
         expected = pd.array([pd.Timedelta(0), pd.Timedelta(2), pd.NaT])
